move AllowTokenPostRequestValidator to its own class
