+++ JabbR/JabbR.csproj
@@ -824,6 +824,7 @@
   <ItemGroup>
     <Compile Include="App_Start\FederatedLogin.Hooks.cs" />
     <Compile Include="App_Start\FederatedLogin.cs" />
+    <Compile Include="Auth\AllowTokenPostRequestValidator.cs" />
     <Compile Include="Commands\WelcomeCommand.cs" />
     <Compile Include="Commands\UpdateCommand.cs" />
     <Compile Include="ContentProviders\SpotifyContentProvider.cs" />
