+++ JabbR/Web.config
@@ -26,7 +26,7 @@
     <add key="googleAnalytics" value="" />
     <!-- populate auth.appId with your janrain appId -->
     <add key="auth.appId" value="" />
-    
+
     <!-- federated identity configuration for enterprise sso (ADFS, SiteMinder, Ping, Windows Azure Active Directory, etc.) -->
     <add key="fedauth.identityProviderUrl" value="" />
     <add key="fedauth.realm" value="" />
@@ -36,7 +36,6 @@
               true = will show the modal identity provider selector, false = will redirect to WAAD (uncomment authorization section as well) -->
     <add key="fedauth.waad.selectorEnabled" value="false" />
     <add key="fedauth.waad.namespace" value="" />
-
   </appSettings>
   <elmah>
     <errorFilter>
@@ -50,7 +49,7 @@
     <pages validateRequest="false" />
     <customErrors mode="Off" />
     <compilation debug="true" targetFramework="4.0" />
-    <httpRuntime requestValidationMode="2.0" requestValidationType="JabbR.App_Start.FederatedLogin+AllowTokenPostRequestValidator" />
+    <httpRuntime requestValidationMode="2.0" requestValidationType="JabbR.Auth.AllowTokenPostRequestValidator" />
     <httpModules>
       <add name="ErrorLog" type="Elmah.ErrorLogModule, Elmah" />
       <add name="ErrorMail" type="Elmah.ErrorMailModule, Elmah" />
